TUC-131 new class GeneralExtractor to remove some code from ComputingComponent
The GeneralExtractor embeds some of the extractions previously in
ComputingComponent. It offers better testability and reduces the size
of computingcomponent.cpp
Also modified the intake extractor and IntakeEvent. Previously the
IntakeEvent constructor was modifying the number of points to make it odd.
It is now the responsibility of the caller. The tests have been modified
according to this new behavior.

	modified:   ../../make/qtcreator/tucucore.pri
		* Adding generalextractor.*
	modified:   ../../src/tucucore/computingcomponent.cpp
	modified:   ../../src/tucucore/computingcomponent.h
		* Simplified, as it now calls GeneralExtractor
	modified:   ../../src/tucucore/drugfilevalidator.cpp
		* Very small modif
	new file:   ../../src/tucucore/generalextractor.cpp
	new file:   ../../src/tucucore/generalextractor.h
		* New class used by ComputingComponent
	modified:   ../../src/tucucore/intakeevent.h
		* Modified the constructor
	modified:   ../../src/tucucore/intakeextractor.cpp
		* Calculates really the number of points for each intake
	modified:   ../../src/tucucore/montecarlopercentilecalculator.cpp
		* Small modif
	modified:   ../../src/tucucore/parameter.cpp
		* Small modif
	modified:   ../tucucore/test_computingcomponentadjustments.h
		* Less compiler messages thanks to static_cast
	modified:   ../tucucore/test_covariateextractor.h
		* Less compiler messages thanks to static_cast
	modified:   ../tucucore/test_intakeextractor.h
		* Modified to fit the new IntakeEvent constructor

+++ test/tucucore/test_covariateextractor.h
@@ -748,22 +748,22 @@ struct TestCovariateExtractor : public fructose::test_base<TestCovariateExtracto
 
             fructose_assert(refreshMap.size() == 403);
             // 17.08.2017 @ 14h00 (start)
-            fructose_assert_eq(refreshMap[DATE_TIME_NO_VAR(2017, 8, 17, 14, 0, 0)].size(), 3);
+            fructose_assert_eq(refreshMap[DATE_TIME_NO_VAR(2017, 8, 17, 14, 0, 0)].size(), static_cast<size_t>(3));
             CHECK_REFRESH(AgeDays, DATE_TIME_NO_VAR(2017, 8, 17, 14, 0, 0), refreshMap);
             CHECK_REFRESH(AgeMonths, DATE_TIME_NO_VAR(2017, 8, 17, 14, 0, 0), refreshMap);
             CHECK_REFRESH(AgeYears, DATE_TIME_NO_VAR(2017, 8, 17, 14, 0, 0), refreshMap);
             // 17.08.2017 @ 14h15
-            fructose_assert_eq(refreshMap[DATE_TIME_NO_VAR(2017, 8, 17, 14, 15, 0)].size(), 1);
+            fructose_assert_eq(refreshMap[DATE_TIME_NO_VAR(2017, 8, 17, 14, 15, 0)].size(), static_cast<size_t>(1));
             CHECK_REFRESH(AgeDays, DATE_TIME_NO_VAR(2017, 8, 17, 14, 15, 0), refreshMap);
 
             // 13.09.2017
-            fructose_assert_eq(refreshMap[DATE_TIME_NO_VAR(2017, 9, 13, 14, 15, 0)].size(), 2);
+            fructose_assert_eq(refreshMap[DATE_TIME_NO_VAR(2017, 9, 13, 14, 15, 0)].size(), static_cast<size_t>(2));
             CHECK_REFRESH(AgeDays, DATE_TIME_NO_VAR(2017, 9, 13, 14, 15, 0), refreshMap);
             CHECK_REFRESH(AgeMonths, DATE_TIME_NO_VAR(2017, 9, 13, 14, 15, 0), refreshMap);
 
             // 13.12.2017
-            fructose_assert_eq(refreshMap.count(DATE_TIME_NO_VAR(2017, 12, 13, 14, 0, 0)), 0);
-            fructose_assert_eq(refreshMap[DATE_TIME_NO_VAR(2017, 12, 13, 14, 15, 0)].size(), 3);
+            fructose_assert_eq(refreshMap.count(DATE_TIME_NO_VAR(2017, 12, 13, 14, 0, 0)), static_cast<size_t>(0));
+            fructose_assert_eq(refreshMap[DATE_TIME_NO_VAR(2017, 12, 13, 14, 15, 0)].size(), static_cast<size_t>(3));
             CHECK_REFRESH(AgeDays, DATE_TIME_NO_VAR(2017, 12, 13, 14, 15, 0), refreshMap);
             CHECK_REFRESH(AgeMonths, DATE_TIME_NO_VAR(2017, 12, 13, 14, 15, 0), refreshMap);
             CHECK_REFRESH(AgeYears, DATE_TIME_NO_VAR(2017, 12, 13, 14, 15, 0), refreshMap);
